# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Modak Challenge API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/notifications/send": {
            "post": {
                "description": "Sends a notification to a user respecting per-type rate limits",
                "tags": [
                    "notifications"
                ],
                "summary": "Send a notification",
                "parameters": [
                    {
                        "description": "Notification payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notification.SendNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/notification.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notification.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "notification.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "notification.SendNotificationRequest": {
            "type": "object",
            "required": [
                "message",
                "type",
                "user_id"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "notification.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}